refactor: flatten base-either and base-maybe into their concrete files
Remove base-either.ts and base-maybe.ts. EitherMatch/EitherOn are now
defined in either.ts; MaybeMatch/MaybeOn in maybe.ts. Left/Right extend
Either directly, Just/Nothing extend Maybe directly.
